refactor: flatten the eyreish module into the crate root
`eyreish` was miette's vendored fork of eyre/anyhow, walled off in its own
subdirectory. Dissolve it into top-level modules so the error-report machinery
reads as part of the crate:

  eyreish/mod.rs                    -> report.rs      (Report, set_hook, ReportHandler)
  eyreish/error.rs                  -> report_impl.rs (ErrorImpl + vtable)
  eyreish/ptr.rs                    -> ptr.rs
  eyreish/kind.rs                   -> kind.rs
  eyreish/macros.rs                 -> macros.rs
  eyreish/into_diagnostic.rs        -> into_diagnostic.rs
  eyreish/fmt.rs                    -> fmt.rs
  eyreish/context.rs + wrapper.rs   -> wrap_err.rs (with the WrapErr trait)

The unsafe vtable core moves verbatim. Cross-module `super::` paths become
`crate::`, and items formerly reached as a child module (`Report.inner`,
`capture_handler`) become `pub(crate)`. The module-wide clippy allows that
eyreish carried are re-scoped to the files that need them.

The public API is unchanged: every name still re-exports at the crate root
(`miette::Report`, `WrapErr`, `Context`, `miette!`, `IntoDiagnostic`, ...).
